BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦 BING CHILLING 🥶 🍦

+++ app/src/main/java/com/example/demo_2340/GameScreen1.java
@@ -168,7 +168,7 @@ public class GameScreen1 extends AppCompatActivity {
 
         rootView.invalidate();
     }
-    private void moveEnemySprite(int deltaX, int deltaY) {
+    private void moveEnemySprite(double deltaX, double deltaY) {
         FrameLayout fr = findViewById(R.id.playerInfoView);
 
         double newX = spriteEnemy.getxPosition();
